requirements.txt
flask
flask-session
google-api-python-client
google-auth
google-auth-oauthlib
google-cloud-aiplatform
google-cloud-secret-manager
python-dotenv
google-generativeai